Refactored SkunkController/Game class. Move responsibilities to Game.
