+++ SkunkProject/src/skunk/domain/Game.java
@@ -1,12 +1,30 @@
 package skunk.domain;
 
-import java.util.ArrayList;
+//import java.util.ArrayList;
 
 public class Game {
 	
 	private boolean StartGame = false;
-	private boolean EndGame = false;
-	private ArrayList<Player> totalPlayers = new ArrayList();
+//	private boolean EndGame = false;
+//	private ArrayList<Player> totalPlayers = new ArrayList();
+	
+	private int iNumOfPlayers;
+	private String[] NameOfPlayers;
+	
+	//Constants
+	private static final int CONSTANT_MAX_NUMBER_PLAYERS = 30;
+	private static final int ERROR_INVALID_PLAYER_NUMBER = -1;
+	
+	//Class objects
+	private SkunkUI ui;
+	private Turn turn;
+	
+	// Game Constructor 
+	public Game() {
+		ui = new SkunkUI();
+		turn = new Turn(ui);
+		this.NameOfPlayers = new String[CONSTANT_MAX_NUMBER_PLAYERS];
+	}
 	
 	public String getGameStatus()
 	{
@@ -26,34 +44,116 @@ public class Game {
 		this.StartGame = true;
 	}
 	
-	public boolean endGame()
-	{
-		return this.EndGame = false;
-	}
+//	public boolean endGame()
+//	{
+//		return this.EndGame = false;
+//	}
+//	
+//	
+//	// Return total players.
+//	public int getTotalPlayers()
+//	{
+//		int total = totalPlayers.size();
+//		return total;
+//		
+//	}
+//
+//	// Method to add Player to Game. 
+//	public void addPlayer(String string) 
+//	
+//	{
+//		Player player = new Player(string);
+//		totalPlayers.add(player);
+//		
+//	}
+//	
+//	// Method to REMOVE all Players from Game. 
+//
+//	public void removePlayers() 
+//	{
+//		totalPlayers.clear();
+//	}
+//
+	public boolean run() 
 	
-	
-	// Return total players.
-	public int getTotalPlayers()
 	{
-		int total = totalPlayers.size();
-		return total;
+		// Ask for number of players.
+		askAndParse_NumberOfPlayers();
+		
+		// If the number of players entered in invalid, try again.
+		int iStatus = getNumberOfPlayers(this.iNumOfPlayers);
+		while( iStatus < 0 ) 
+		{
+			askAndParse_NumberOfPlayers();
+			iStatus = getNumberOfPlayers( this.iNumOfPlayers );
+		} 
+		
+		// Ask for player's name, can add try and catch here.
+		savePlayerNamesInArray();
+		
+		
+		//For P1.2: Just one Player.
+		//One complete interactive turn of skunk with one human player.
+		
+		iStatus = turn.playTurn();
+		
+		ui.printLine( "Turn is OVER." );
+		
+		return true;
 		
 	}
-
-	// Method to add Player to Game. 
-	public void addPlayer(String string) 
 	
+	public void askAndParse_NumberOfPlayers() 
 	{
-		Player player = new Player(string);
-		totalPlayers.add(player);
-		
+		String strTemp;
+		ui.printLine( "*****" );
+		try
+		{
+			strTemp = ui.printLineReadResponse( "Enter number of players?" );
+			this.iNumOfPlayers = Integer.parseInt( strTemp );
+		}
+		catch( Exception e )
+		{
+			//Todo: if string entered is other than integer.
+		}
+	}
+
+	//**********************************************************
+	// SkunkController will communicate with UI and get the number of players.
+	// Set the parsed number of players in the Player class.
+	//**********************************************************
+	
+	public int getNumberOfPlayers(int iNumOfPlayers)
+	{
+		if( iNumOfPlayers <= 0 )
+		{
+			ui.printLine( "invalid iNumOfPlayers: " + iNumOfPlayers + "\nEnter number of players greater than 0?\n" );
+			return ERROR_INVALID_PLAYER_NUMBER;
+		}
+
+		return 0;
 	}
 	
-	// Method to REMOVE all Players from Game. 
-
-	public void removePlayers() 
+	//**********************************************************
+	// SkunkController will communicate with UI and get the names of players.
+	// Set the names of the players in the Player class.
+	//**********************************************************
+		
+	public void savePlayerNamesInArray() 
 	{
-		totalPlayers.clear();
+		String strTemp;
+		for ( int iii =0; iii < this.iNumOfPlayers; iii++ )
+		{
+			try
+			{
+				strTemp = ui.printLineReadResponse( "\nEnter Name of Player " + (iii+1) + ": " );
+				this.NameOfPlayers[iii] = strTemp;
+			}
+			catch( Exception e )
+			{
+				//ToDo: empty string entered or just return key hit.
+			}
+		}
 	}
 
 }
