+++ SkunkProject/src/skunk/domain/SkunkApp.java
@@ -1,4 +1,3 @@
-// Added in comment for developer branch
 package skunk.domain;
 import edu.princeton.cs.introcs.StdOut;
